requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
google-cloud-vision==3.4.4
chromadb==1.0.0
langchain==0.0.331
langchain-community==0.0.1
ollama==0.2.1
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.7.0
huggingface-hub==0.20.3
bcrypt==4.0.1
email-validator==2.0.0
